requirements/airflow.txt
apache-airflow-providers-docker
apache-airflow-providers-apache-spark
apache-airflow-providers-amazon

pyspark==3.5.1
python-dotenv==1.0.1
